+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [13.0.1] (IN PROGRESS)
 
 * Missing values in the version history modal for Suppressed from discovery and Staff suppressed fields. Fixed UIIN-3277.
+* ECS FOLIO instances: Inventory version history feature toggle/icon. Refs UIIN-3284.
 
 ## [13.0.0](https://github.com/folio-org/ui-inventory/tree/v13.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.12...v13.0.0)